corretto componente livewire e inserite categorie, sortate in ordine alfabetico, all'interndo del form. Ho iniziato a gestire gli errori e la validazione del componente livewire, bisogna gestire il tempo di risposta, ancora da fare il frontend, per il resto funziona tutto

+++ database/migrations/2023_05_22_153251_create_announcements_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
 
             $table->string('title');
             $table->string('place');
-            $table->float('price');
+            $table->decimal('price',9,2);
             $table->text('description');
             
             $table->timestamps();
